refactor: rename github addon
